Organize utility modules under src/utils
