+++ src/utils/inventoryUtils.js
