+++ tests/AppCalendarFilter.test.jsx
@@ -1,17 +1,17 @@
 /* eslint-env jest */
 import { render, screen, fireEvent } from '@testing-library/react';
 // Mock asset imports used by App
-jest.mock('../assets/conceptB-ETF-Life-dark.svg', () => 'data:image/svg+xml;base64,PHN2Zy8+');
-jest.mock('../assets/conceptB-ETF-Life-light.svg', () => 'data:image/svg+xml;base64,PHN2Zy8+');
+jest.mock('../src/assets/conceptB-ETF-Life-dark.svg', () => 'data:image/svg+xml;base64,PHN2Zy8+');
+jest.mock('../src/assets/conceptB-ETF-Life-light.svg', () => 'data:image/svg+xml;base64,PHN2Zy8+');
 
-jest.mock('../api', () => ({
+jest.mock('../src/api', () => ({
   fetchWithCache: jest.fn(() => Promise.resolve({ data: [], cacheStatus: 'fresh', timestamp: '' })),
   clearCache: jest.fn(),
 }));
 
-jest.mock('../config', () => ({ API_HOST: '' }));
+jest.mock('../src/config', () => ({ API_HOST: '' }));
 
-import App from '../App';
+import App from '../src/App';
 
 beforeAll(() => {
   globalThis.fetch = jest.fn(() => Promise.resolve({}));
