+++ tests/DisclaimerTab.test.jsx
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 import { render, screen } from '@testing-library/react';
-import DisclaimerTab from '../DisclaimerTab';
+import DisclaimerTab from '../src/DisclaimerTab';
 
 test('renders disclaimer heading', () => {
   render(<DisclaimerTab />);
