+++ tests/FaqTab.test.jsx
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 import { render, screen } from '@testing-library/react';
-import FaqTab from '../FaqTab';
+import FaqTab from '../src/FaqTab';
 
 test('renders faq heading', () => {
   render(<FaqTab />);
