+++ tests/dividendUtils.test.js
@@ -1,4 +1,4 @@
-import { getTomorrowDividendAlerts } from '../dividendUtils';
+import { getTomorrowDividendAlerts } from '../src/utils/dividendUtils';
 
 describe('getTomorrowDividendAlerts', () => {
   test('detects ex-dividend and payment events for tomorrow', () => {
